Rebase onto current main and keep the size gates satisfied (Fixes #522)
Rebasing onto the action-registry work required reconciling the keymap-aware
startup and test context, the renamed launch-signature helper, and the upstream
fix for the trusted-probe invocation expectation, which now supersedes the one
this branch carried.

messages.rs and state/events.rs are both at the 1000-line limit on main, so the
observation additions pushed them over. PullRequestsMessage moves into
src/messages/prs.rs alongside the existing extracted message modules, and the
JSP host startup moves into a named helper so run_tui stays within the function
length limit. No limit was raised and no suppression was added.
